rename whitelist_generator -> whitelist
